Clean up JSONPropertyTest
